refactor(pull-request): move pr list functionality into separate file
