+++ .vscode/launch.json
@@ -20,8 +20,7 @@
             "env": {
                 "NODE_ENV": "testing",
                 "GH_USER": "protoEvangelion",
-                "GH_TOKEN": "0001",
-                "GENERATE_NEW_TOKEN": "true"
+                "GH_TOKEN": "0001"
             },
             "osx": {
                 "DEBUG": "nock.*"
